+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.5.0]
+
+* Features:
+  Add call back [onSlidingPage] when is sliding page, you can change other widgets state in page.[ExtendedImageSlidePage]
+* Add [enableSlideOutPage] parameter to define whether enable slide out page. [ExtendedImage]  
+
 ## [0.4.3]
 
 * Breaking Change: